requirements.txt
streamlit>=1.33
pandas
alpaca-py
python-dotenv